+++ Mummy.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class Mummy extends Enemy {
     public static final int ATTKOFFSET = 35;
-    public Mummy(double x, double y) {
+    public Mummy(int x, int y) {
         super("Mummy", x, y, 75, 100, 0, 750, 0, 0, new AttackStats(5, 30, 12, 400, 1, new Color(145, 104, 0)));
     }
