+++ Wight.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class Wight extends Enemy {
     public static final int ATTKOFFSET = 35;
-    public Wight(double x, double y) {
+    public Wight(int x, int y) {
         super("Wight", x, y, 75, 80, 5, 1500, 450, 1000, new AttackStats(8, 35, 13, 300, -1, new Color(80, 80, 80)));
     }
